temp: x11: Keep track of variable refresh rate support in X windows
Use the _VARIABLE_REFRESH property to identify top-level X windows
that support variable refresh rate.

This has the limitation that only top-level windows are tracked, and
is only intended as a temporary measure for the purpsoe of testing.

A better solution for this would be to have Xwayland inform the
Wayland compositor about surfaces that support variable refresh
rate via a Wayland protocol.

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1154