temp: wayland/surface: flip committed client frames as soon as possible
This synchronizes page flipping with client surface commits, allowing
client frames to be flipped as soon as possible.

Flipping at the same frequency as the VRR client is drawing its frames
is required in order to communicate the requested refresh rate to the
VRR driver.

This is generally a bad idea because it disables frame scheduling
completely, introducing output latency in multi-window scenarios where
multiple clients commit their frames. It also introduces jitter when
those clients commit their frames at different frequencies.

A better solution would be to enable this synchronization only when
VRR is enabled for the output, so that the VRR client is the only
client being drawn on the monitor. However, this would require frame
scheduling to be performed per-output, which is still not the case in
Mutter.

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1154