clutter/stage: Add API to check if a stage window can redraw
A stage window can redraw only when it has a framebuffer available to
be used as the back buffer.

The stage window is double-buffered, so it only has a maximum of one
framebuffer available to be used as the back buffer at any given time.

If that one framebuffer was submitted for swap, the stage window is
considered not ready to redraw.

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1241