backend/native: Allow stage updates to occur without stage redraws
A stage update includes moving the hardware cursor and dispatching
input events to Wayland clients, which are operations that shouldn't
be throttled to stage redraws since they need to be performed at
least once in every refresh period.

Instead, allow stage updates to occur at their scheduled update time,
and rely on the fact that stage updates would only perform the
operations that were queued and can be accomplished for the udpdate,
including redraws.

Since updates are still scheduled to occur within the next refresh
period whenever they were scheduled, the update frequency in idle and
light rendering workloads isn't affected by this change and remains
a maximum of once per refresh period, and only when an update is due.

However, in scenarios where client rendering is more intense and
Mutter is expected to consistently miss the next vblank due to
implicit synchronization on client buffers and general high load on
the GPU pipeline, this change guarantees that an update would still
occur once per refresh period.

The result of this is smooth hardware cursor movement even under load,
and a maximum input event latency of a single refresh period between
Mutter and Wayland clients.

Since this change does not affect the rate of updates in idle and
light rendering workloads, it should provide a good balance between
latency and power usage.

A similar change can't be done for the X11 backend yet since it cannot
currently guarantee that updates would usually occur at a consistent
rate of once in every refresh period under high load.

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1241